requirements.txt
Flask==2.2.5
openai==0.28.1
Pillow==9.5.0
Werkzeug==2.2.3
gunicorn==20.1.0
requests==2.31.0 